move: localStorageEffect 파일 store 폴더 내부로 변경
